Sketch a partial ast and generator
The structure of the ast is largelly inspired by Haxe's haxe.macro.Expr,
since that has proven to be incredebly extensible and, at the sime, easy
to traverse.  Loads of features still missing though...

However, these definitions aren't of much use yet.  I first need to
document the ast types; specifically, it's important to define what each
:String can/should be. (maybe I should use abstracts instead of
plain Strings)

Also, I still expect to make some drastic changes to the ast.  In
particular, I'm starting to think it will be easier to traverse and
transform the ast if its HLists and VLists are defined as recursive
HDefs and VDefs, as opposed to the current independent implementation.
This change would keep the ast flatter and simpler, and we wouldn't have
to iterate lists with only one element.  For example, most sections
would be represented by
    { expr : VSection("some-label", { expr : HText("Some title"), [_]), pos : _ }, pos : _ }
instead of
    { expr : VSection("some-label", [ { expr : [HText("Some title"), [_])], pos : _ } ], pos : _ }

Oh, and are you ok with leaving the automatic label generation to the
parser?

Fyi, because of HTML specs and implementations, labels can only have
a-zA-Z0-9-_. and must begin with a letter; and since I don't like _ and
I already want . as a label path separator in the generation phase –
chapter-label.section-label.subsection-label – that pretty much only
leaves letters and -.

Finally, let me know if you have any suggestions so far.