Recursive expr lists, some docs and other (mostly minor) changes
The Expr<GenDef> expressions with normalized labels and numbered
elements, mentioned in the AST docs, are still something to do.

Note: it's unnecessary to create intermediate DOM-like objects during
HTML generation; on the other hand, it probably makes sense to use the
dom4 library plus some macros for testing the output of the generator,
instead of relying on string comparison or matching for that.