Gen: [lex] change math from $...$ to $$...$$; remove $$$... syntax
Obviously, also fix the guide.

This "is because dollar signs appear too often in non-mathematical
settings, which could cause some text to be treated as mathematics
unexpectedly. For example, with single-dollar delimiters, ”... the cost
is $2.50 for the first one, and $2.00 for each additional one ...” would
cause the phrase “2.50 for the first one, and” to be treated as
mathematics since it falls between dollar signs."
—MathJax: TeX and LaTeX math delimiters