Gen: [nt] reimplement trimming of horizontal elements
Instead of converting to and back from a linearised representation, we
solve the problem with two depth first searches, keeping only a common
reference to whether the previous element was or was not a wordspace.

The trimming does not remove any elements from the tree; we replace any
redundant wordspaces by `HEmpty` elements, that will later get purged by
a new `hclean` function.