Server: assume nodejs-legacy was installed
There's a whole discussion whether or not the `node` name is correct or
not from the point of view of a linux distribution.

Let's dodge that and simply use `nodejs-legacy` for a `node` alias to
the package default of `nodejs`, because the first is much more common
in the Node.js world.  (even though the concerns of Debian and Ubuntu
developers do make _some_ sence)