manu: remove all vertical markdown syntax
This syntax is hardly ever used (only in quotations), is much less
regular and generates far more cryptic errors.

After some discussions, we've decided to remove support for `>foo @bar`
quotations, `# foo` headings and `#FIG#` figures.

Finally, this is implemented as a major version change and provides no
compatibility or warnings for the use of the old/unsupported syntax.