Use no-cap token to prevent LFS rate limits [server]
While we discuss with GitHub's support how to best handle #148 (and if,
perhaps, there's an issue with the rate limit rules for the LFS api),
this will serve as a workaround.

Basically, we use a token without any real capabilities merely as a mean
to identify ourselves to GitHub and get a higher rate limit.

*This assumes a responsible configuration where the NO_CAP_GH_TOKEN
token really lacks all capabilities.*