Report \row position if no. of columns differ [manu]
Fixes #173.

It's not very practical to report the position of the entire row, as
opposed to just the `\row` command that starts it.  This would require a
different AST or, more simply, a very ugly conditional within an
assertion.

Maybe we can improve this further if (when) we upgrade the error from an
assertion to a proper `parser.ParserError`.  However, that would
probably mean to change the AST: position reporting for proper errors
should _not_ be entangled with the error generation or storage.