Update mathjax-node and prevent issue with nested dependency
mathjax-node's nested dependency to mathjax ended up being automatically
updated to v2.7.4, which introduces the issue of never completing
typesetting attempts of invalid TeX with unclosed brackets.

While the issue only seemed to affect a unit test case, it is an
important aspect of math syntax validation that, in our opinion,
justifies the upgrade.